Fix jump-to-top bug when removing trailing spaces.
This bug was due to the fact that changing the editor's text
internally (by calling `QsciScintilla::setText()`) causes the
editting buffer to be reset.  To fix the issue, all that's needed
is to save the cursor and scrollbar positions before updating the
text and then reseting them after the change is complete.